# Improved Config performance