--- a/src/core/Akka/Configuration/Hocon/HoconValue.cs
+++ b/src/core/Akka/Configuration/Hocon/HoconValue.cs
@@ -20,6 +20,9 @@ namespace Akka.Configuration.Hocon
     /// </summary>
     public class HoconValue : IMightBeAHoconObject
     {
+        private static readonly Regex EscapeRegex = new Regex("[ \t:]{1}", RegexOptions.Compiled);
+        private static readonly Regex TimeSpanRegex = new Regex(@"^(?<value>([0-9]+(\.[0-9]+)?))\s*(?<unit>(nanoseconds|nanosecond|nanos|nano|ns|microseconds|microsecond|micros|micro|us|milliseconds|millisecond|millis|milli|ms|seconds|second|s|minutes|minute|m|hours|hour|h|days|day|d))$", RegexOptions.Compiled);
+
         /// <summary>
         /// Initializes a new instance of the <see cref="HoconValue"/> class.
         /// </summary>
@@ -146,7 +149,7 @@ namespace Akka.Configuration.Hocon
         /// </returns>
         public bool IsString()
         {
-            return Values.Any() && Values.All(v => v.IsString());
+            return Values.Count > 0 && Values.All(v => v.IsString());
         }
 
         private string ConcatString()
@@ -192,7 +195,6 @@ namespace Akka.Configuration.Hocon
                     return false;
                 default:
                     throw new NotSupportedException($"Unknown boolean format: {v}");
-
             }
         }
 
@@ -372,7 +374,7 @@ namespace Akka.Configuration.Hocon
         {
             string res = GetString();
 
-            var match = Regex.Match(res, @"^(?<value>([0-9]+(\.[0-9]+)?))\s*(?<unit>(nanoseconds|nanosecond|nanos|nano|ns|microseconds|microsecond|micros|micro|us|milliseconds|millisecond|millis|milli|ms|seconds|second|s|minutes|minute|m|hours|hour|h|days|day|d))$");
+            var match = TimeSpanRegex.Match(res);
             if (match.Success) 
             {
                 var u = match.Groups["unit"].Value;
@@ -490,13 +492,11 @@ namespace Akka.Configuration.Hocon
             return "<<unknown value>>";
         }
 
-        private static readonly Regex Regexp = new Regex("[ \t:]{1}", RegexOptions.Compiled);
-
         private string QuoteIfNeeded(string text)
         {
             if (text == null) return "";
 
-            if (Regexp.IsMatch(text))
+            if (EscapeRegex.IsMatch(text))
             {
                 return "\"" + text + "\"";
             }